wg-hassio-alarmme/rootfs/requirements.txt
aiohttp>=3.9.0
paho-mqtt>=1.6.0

